Fix Component to be stored properly
* fix PreparedStatementBuilder to handle null properly
* fix Component to convert 'hosts' to List<String> <-> JSON
* fix table create script to let port be INTEGER, not BIGINT

+++ storage/core/src/main/resources/mysql/create_tables.sql
@@ -273,7 +273,7 @@ CREATE TABLE IF NOT EXISTS components (
   name VARCHAR(256) NOT NULL,
   hosts TEXT NOT NULL,
   protocol VARCHAR(256),
-  port BIGINT,
+  port INTEGER,
   timestamp BIGINT,
   PRIMARY KEY (id)
 );
